v1

+++ articles/views.py
@@ -1,22 +1,30 @@
+from http import server
+from msilib.schema import ServiceInstall
 from django.shortcuts import get_object_or_404
+from django.db.models import Count
 
 from rest_framework import status
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
-
+from rest_framework.permissions import IsAuthenticated
 from .models import Article, Comment
-from .serializers import ArticleListSerializer, ArticleSerializer, CommentSerializer
+from .serializers.article import ArticleListSerializer, ArticleSerializer
+from .serializers.comment import CommentSerializer
 
 
 @api_view(['GET', 'POST'])
 def article_list_or_create(request):
 
     def article_list():
-        articles = Article.objects.all()
+        # comment 개수 추가
+        articles = Article.objects.annotate(
+            comment_count=Count('comments', distinct=True),
+            like_count=Count('like_users', distinct=True)
+        )
         serializer = ArticleListSerializer(articles, many=True)
         return Response(serializer.data)
-
-    def create_article():
+    
+    def create_article(request):
         serializer = ArticleSerializer(data=request.data)
         if serializer.is_valid(raise_exception=True):
             serializer.save(user=request.user)
@@ -25,7 +33,7 @@ def article_list_or_create(request):
     if request.method == 'GET':
         return article_list()
     elif request.method == 'POST':
-        return create_article()
+        return create_article(request)
 
 
 @api_view(['GET', 'PUT', 'DELETE'])
@@ -37,42 +45,54 @@ def article_detail_or_update_or_delete(request, article_pk):
         return Response(serializer.data)
 
     def update_article():
-        serializer = ArticleSerializer(instance=article, data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data)
+        if request.user == article.user:
+            serializer = ArticleSerializer(instance=article, data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data)
 
     def delete_article():
-        article.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user == article.user:
+            article.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     if request.method == 'GET':
         return article_detail()
     elif request.method == 'PUT':
-        return update_article()
+        if request.user == article.user:
+            return update_article()
     elif request.method == 'DELETE':
-        return delete_article()
+        if request.user == article.user:
+            return delete_article()
 
 
 @api_view(['POST'])
 def like_article(request, article_pk):
-    article = get_object_or_404(Article, article_pk)
+    article = get_object_or_404(Article, pk=article_pk)
     user = request.user
-    if article.like_users.all().exists(user=user):
+    if article.like_users.filter(pk=user.pk).exists():
         article.like_users.remove(user)
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        serializer = ArticleSerializer(article)
+        return Response(serializer.data)
     else:
         article.like_users.add(user)
-        return Response(status=status.HTTP_201_CREATED)
+        serializer = ArticleSerializer(article)
+        return Response(serializer.data)
 
 
 @api_view(['POST'])
 def create_comment(request, article_pk):
+    user = request.user
     article = get_object_or_404(Article, pk=article_pk)
-
+    
     serializer = CommentSerializer(data=request.data)
     if serializer.is_valid(raise_exception=True):
-        serializer.save(article=article)
+        serializer.save(article=article, user=user)
+
+        # 기존 serializer 가 return 되면, 단일 comment 만 응답으로 받게됨.
+        # 사용자가 댓글을 입력하는 사이에 업데이트된 comment 확인 불가 => 업데이트된 전체 목록 return 
+        comments = article.comments.all()
+        serializer = CommentSerializer(comments, many=True)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
@@ -82,15 +102,21 @@ def comment_update_or_delete(request, article_pk, comment_pk):
     comment = get_object_or_404(Comment, pk=comment_pk)
 
     def update_comment():
-        serializer = CommentSerializer(instance=comment, data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data)
+        if request.user == comment.user:
+            serializer = CommentSerializer(instance=comment, data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                comments = article.comments.all()
+                serializer = CommentSerializer(comments, many=True)
+                return Response(serializer.data)
 
     def delete_comment():
-        comment.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
+        if request.user == comment.user:
+            comment.delete()
+            comments = article.comments.all()
+            serializer = CommentSerializer(comments, many=True)
+            return Response(serializer.data)
+    
     if request.method == 'PUT':
         return update_comment()
     elif request.method == 'DELETE':
